Moved lua util functions to utils.lua and add it to the lua path
