+++ c2rust-refactor/scripts/utils.lua
@@ -1,0 +1,11 @@
+DEBUG = false
+
+function debug(str)
+    if DEBUG then
+        print(str)
+    end
+end
+
+function starts_with(str, start)
+    return str:sub(1, #start) == start
+end
